+++ etherparse/src/internet/mod.rs
@@ -10,4 +10,5 @@ pub mod ipv6_fragment_header;
 pub mod ipv6_fragment_header_slice;
 pub mod ipv6_header;
 pub mod ipv6_header_slice;
-pub mod ipv6_raw_extension;
+pub mod ipv6_raw_ext_header;
+pub mod ipv6_raw_ext_header_slice;